// src/resources/smaHandler/sip-media-application.ts
export enum ActionTypes {
  CALL_AND_BRIDGE = 'CallAndBridge',
  HANGUP = 'Hangup',
  JOIN_CHIME_MEETING = 'JoinChimeMeeting',
  MODIFY_CHIME_MEETING_ATTENDEES = 'ModifyChimeMeetingAttendees',
  PAUSE = 'Pause',
  PLAY_AUDIO = 'PlayAudio',
  PLAY_AUDIO_AND_GET_DIGITS = 'PlayAudioAndGetDigits',
  RECEIVE_DIGITS = 'ReceiveDigits',
  RECORD_AUDIO = 'RecordAudio',
  SEND_DIGITS = 'SendDigits',
  SPEAK = 'Speak',
  SPEAK_AND_GET_DIGITS = 'SpeakAndGetDigits',
  START_BOT_CONVERSATION = 'StartBotConversation',
  VOICE_FOCUS = 'VoiceFocus',
  START_CALL_RECORDING = 'StartCallRecording',
  STOP_CALL_RECORDING = 'StopCallRecording',
  PAUSE_CALL_RECORDING = 'PauseCallRecording',
  RESUME_CALL_RECORDING = 'ResumeCallRecording',
}
export enum BridgeEndpointType {
  AWS = 'AWS',
  PSTN = 'PSTN',
}
export interface CallAndBridgeEndpoints {
  BridgeEndpointType: BridgeEndpointType;
  Uri: string;
  Arn?: string;
}

export interface AudioSource {
  Type: 'S3';
  BucketName: string;
  Key: string;
}

export interface CommonActionParameters {
  CallId?: string;
  ParticipantTag?: ParticipantTag;
}
export interface CallAndBridgeActionParameters {
  CallTimeoutSeconds: number;
  CallerIdNumber: string;
  RingbackTone?: AudioSource;
  Endpoints: CallAndBridgeEndpoints[];
  SipHeaders?: { [key: string]: string };
}
// ...
export enum PollyVoiceIds {
  ADITI = 'Aditi',
  AMY = 'Amy',
  ASTRID = 'Astrid',
  BIANA = 'Bianca',
  BRIAN = 'Brian',
  CAMILA = 'Camila',
  CARLA = 'Carla',
  CARMEN = 'Carmen',
  CELINE = 'Celine',
  CHANTAL = 'Chantal',
  CONGITA = 'Conchita',
  CRISTIANO = 'Cristiano',
  DORA = 'Dora',
  EMMA = 'Emma',
  ENRIQUE = 'Enrique',
  EWA = 'Ewa',
  FILIZ = 'Filiz',
  GABRIELLE = 'Gabrielle',
  GERIANT = 'Geraint',
  GIORGI = 'Giorgio',
  GWYNETH = 'Gwyneth',
  HANS = 'Hans',
  INES = 'Ines',
  IVT = 'Ivy',
  JACKEK = 'Jacek',
  JAN = 'Jan',
  JOANNA = 'Joanna',
  JOEY = 'Joey',
  JUSTIN = 'Justin',
  KAJAL = 'Kajal',
  KARL = 'Karl',
  KENDRA = 'Kendra',
  KEVIN = 'Kevin',
  KIMBERLY = 'Kimberly',
  LEA = 'Lea',
  LIV = 'Liv',
  LOTTE = 'Lotte',
  LUCIA = 'Lucia',
  LUPE = 'Lupe',
  MADS = 'Mads',
  MAJA = 'Maja',
  MARLENE = 'Marlene',
  MATHIEU = 'Mathieu',
  MATTHEW = 'Matthew',
  MAXIM = 'Maxim',
  MIA = 'Mia',
  MIGUEL = 'Miguel',
  MIZUKI = 'Mizuki',
  NAJA = 'Naja',
  NICOLE = 'Nicole',
  OLIVIA = 'Olivia',
  PENELOPE = 'Penelope',
  RAVEENA = 'Raveena',
  RICARDO = 'Ricardo',
  RUBEN = 'Ruben',
  RUSSELL = 'Russell',
  SALLI = 'Salli',
  SEOYEON = 'Seoyeon',
  TAKUMI = 'Takumi',
  TATYTANA = 'Tatyana',
  VICKI = 'Vicki',
  VICTORIA = 'Victoria',
  ZEINA = 'Zeina',
  ZHIYU = 'Zhiyu',
  ARIA = 'Aria',
  AYANDA = 'Ayanda',
}

export enum TextType {
  SSML = 'ssml',
  TEXT = 'text',
}
export enum ContentType {
  SSML = 'SSML',
  PLAIN_TEXT = 'PlainText',
}
export enum PollyLanguageCodes {
  ARB = 'arb',
  CMN_CN = 'cmn-CN',
  CY_GB = 'cy-GB',
  DA_DK = 'da-DK',
  DE_DE = 'de-DE',
  EN_AU = 'en-AU',
  EN_GB = 'en-GB',
  EN_GB_WLS = 'en-GB-WLS',
  EN_IN = 'en-IN',
  EN_US = 'en-US',
  ES_ES = 'es-ES',
  ES_MX = 'es-MX',
  ES_US = 'es-US',
  FR_CA = 'fr-CA',
  FR_FR = 'fr-FR',
  IS_IS = 'is-IS',
  IT_IT = 'it-IT',
  JA_JP = 'ja-JP',
  HI_IN = 'hi-IN',
  KO_KR = 'ko-KR',
  NB_NO = 'nb-NO',
  NL_NL = 'nl-NL',
  PL_PL = 'pl-PL',
  PT_BR = 'pt-BR',
  PT_PT = 'pt-PT',
  RO_RO = 'ro-RO',
  RU_RU = 'ru-RU',
  SV_SE = 'sv-SE',
  TR_TR = 'tr-TR',
  EN_NZ = 'en-NZ',
  EN_ZA = 'en-ZA',
}
export enum Engine {
  STANDARD = 'standard',
  NEURAL = 'neural',
}